# backend/src/main/openapi/documentation.yaml
openapi: "3.0.3"
info:
  title: "FinGenius"
  description: "Financial Application (Backend)"
  version: "1.0.0"
servers:
  - url: "/"
paths:
...
  /invoices:
    # Get Invoices
    get:
      description: "Route to get invoice(s)"
      parameters:
        - name: "id"
          in: "query"
          description: "ID of the invoice to retrieve"
          required: false
          schema:
            type: "string"
        - name: "type"
          in: "query"
          description: "Type of the invoice(s) to retrieve"
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
        "404":
          description: "Not Found"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                Example#1:
                  value: "Invoice not found."
                Example#2:
                  value: "Partner not found."
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                Example#1:
                  value: "Failed to retrieve invoice(s)"

    # Add Invoice
    post:
      description: "Route to add an invoice"
      requestBody:
        content:
          '*/*':
            schema:
              $ref: "#/components/schemas/Invoice"
        required: true
      responses:
        "201":
          description: "Created"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                Example#1:
                  value: "Invoice added successfully."
        "500":
          description: "Internal Server Error"
          content:
            '*/*':
              schema:
                type: "string"
              examples:
                Example#1:
                  value: "Failed to add invoice"
          "400":
            description: "Bad Request"
            content:
              '*/*':
                schema:
                  type: "string"
                examples:
                  Example#1:
                    value: "Invalid invoice data format"
...
components:
  schemas:
...
    InvoiceItem:
      type: "object"
      properties:
        productName:
          type: "string"
        quantity:
          type: "string"
        rate:
          type: "string"
        valueOfSupplies:
          type: "string"
        salesTax:
          type: "string"
        netAmount:
          type: "string"
    Invoice:
      type: "object"
      properties:
        _id:
          type: "string"
        type:
          type: "string"
        invoiceDate:
          type: "string"
        dueDate:
          type: "string"
        invoiceNo:
          type: "string"
        partnerId:
          type: "string"
        creditTerm:
          type: "string"
        reference:
          type: "string"
        invoiceTotal:
          type: "string"
        invoiceItems:
          type: "array"
          items:
            $ref: "#/components/schemas/InvoiceItem"
...